toolchain-bootstrap: add project for building an LLVM toolchain
The goal of this project is to build highly portable, generic
toolchains.

This project is a spiritual successor to functionality I built for
python-build-standalone. An immediate goal is to rip out the toolchain
building from that project and have it consume pre-built toolchains
from this project.

I've chosen Earthly to help orchestrate all the building because the
DSL is highly expressive.