apple-sdk: initial commit
This crate is a port of functionality currently in the tugger-apple
crate. I am the sole author of the original code and am relicensing it
under MIT OR Apache-2.0 for compatibility with the broader Rust
ecosystem.

The first implemented feature is Xcode searching. This is a prerequisite
for interacting with Apple SDKs.