toolchain-bootstrap: disable zstd on macOS
It can pick up a zstd from an external source (e.g. Homebrew) until
we force it to use a zstd we build from source. I'm a huge proponent
of zstd. But I don't believe LLVM 15 has much support for zstd yet.
So we're not missing much by excluding it at the moment. We should
definitely add support for zstd into the Linux and macOS toolchains
at some point though!