toolchain-bootstrap: upgrade LLVM 15.0.7 -> 16.0.1
As part of this the standalone bolt build was deleted and bolt is
now built as part of LLVM.

IIRC I wanted to do this originally but held back because a sufficiently
bug-free bolt required commits in unreleased LLVM 16. So path of least
resistance was to build bolt from an unreleased Git commit. Since bolt
now works well in LLVM 16, it should be safe to include bolt in the
distribution.