toolchain-bootstrap: work around VCSRevision.h wonkiness
We aren't building from GIt and we don't need the VCSRevision.h file
to have meaningful content. So disable it.

But wait there's more!

We get build errors due to a missing `VCSRevision.h`. I think the bolt
build is missing a dependency to build this file. This seems to reliably
reproduce if `LLVM_ENABLE_PROJECTS` has `bolt` before `clang`. I guess
the order the projects are defined implicitly turns into build order
[when using ninja]? Adding a forced evaluation of the missing target
appears to make the problem go away.