toolchain-bootstrap: force target triple when building on Apple
This used to "just work." But for reasons unknown to me, GitHub Actions
Apple ARM runners result in `config.guess` thinking it is a x86-64
machine and this causes LLVM's CMake to build x86-64 binaries which
target x86-64 by default!

Our Python code has working target sniffing. So we use that to feed
an appropriate triple into LLVM's CMake.