Port from Earthly to Dockerfile
The Earthly project looks to be dead. In the years since it was created,
BuildKit + Dockerfile evolved to support a lot of the same features. Not
quite all the features. But enough for it to not be too annoying to use
a vanilla Dockerfile.

This commit ports us from Earthly to Docker.